set up types for the frontend, no type checking yet

+++ run/program.c
@@ -1,7 +1,7 @@
 int main()
 {
-	int x;
+	short x;
 
 	x = 0;
-	return x;
+	return (int) x;
 }
